+++ src/main/index.ts
@@ -1,6 +1,6 @@
-export * from './interfaces';
 export * from './decorators';
 export * from './exceptions';
 export * from './handlers';
+export * from './interfaces';
 export * from './rester';
 export * from './utils';