+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - refactor(app): refactor project structure
 - refactor(decorators/injector): refactor injector methods
+- perf(core/rester): use yaml instead of json config
 
 ## 0.12.8 => 0.12.10